requirements.txt
pillow
tensorflow
nltk
joblib
scikit-learn==0.23.2
pandas
numpy


